// src/modules/speech/enums/speech-error-code.enum.ts
export enum SpeechErrorCode {
  MODEL_NOT_CONFIGURED = 'MODEL_NOT_CONFIGURED',
  MODEL_NOT_FOUND = 'MODEL_NOT_FOUND',
  MODEL_INVALID = 'MODEL_INVALID',
  MICROPHONE_NOT_CONFIGURED = 'MICROPHONE_NOT_CONFIGURED',
  MICROPHONE_NOT_FOUND = 'MICROPHONE_NOT_FOUND',
  VOSK_UNAVAILABLE = 'VOSK_UNAVAILABLE',
  VOSK_INITIALIZATION_FAILED = 'VOSK_INITIALIZATION_FAILED',
  CAPTURE_UNAVAILABLE = 'CAPTURE_UNAVAILABLE',
  CAPTURE_FAILED = 'CAPTURE_FAILED',
  PROVIDER_INTERNAL_ERROR = 'PROVIDER_INTERNAL_ERROR',
}